regalloc: split int binop consider_*_j2 per backend arch
Move the prepare_op_int_* family of consider_*_j2 methods from the
shared majit-backend-dynasm/src/regalloc.rs into arch-specific
aarch64/regalloc.rs and x86/regalloc.rs, mirroring upstream
rpython/jit/backend/{aarch64,x86}/regalloc.py.

aarch64 entries call force_allocate_reg(op) for the result (AAPCS64
3-operand Rd, Rn, Rm); x86 entries keep force_result_in_reg(op, lhs)
(2-operand dst,src). The earlier cfg-gated branches inside
consider_int_lshift_j2 and consider_uint_mul_high_j2 are removed in
favour of the file-level split.

Promote perform, tp, const_value on RegAlloc to pub(crate) so the
arch modules' impl<'a> RegAlloc<'a> blocks can call them.
